refactor(article): 重构文章相关接口并新增 ArticleResDto 类- 新增 ArticleResDto 类用于文章列表展示 - 重构文章创建接口，优化代码结构 - 更新文章时间轴接口，使用 ArticleResDto 作为返回类型 - 优化标签创建接口，直接使用 BlogTag 对象 - 移除 BlogTag 类中的冗余代码 - 更新相关 mapper 和 service接口

+++ src/test/java/com/xc/blogbackend/service/impl/BlogCategoryServiceImplTest.java
@@ -12,7 +12,7 @@ class BlogCategoryServiceImplTest {
         BlogCategoryServiceImpl blogCategoryService = new BlogCategoryServiceImpl();
 
         // 创建一个测试用的分类id
-        Integer categoryId = 4;
+        Long categoryId = 4L;
 
         // 创建一个 BlogCategory 对象，用于存储查询结果
         String categoryNameById = blogCategoryService.getCategoryNameById(categoryId);
